requirements.txt
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.0
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector==2.2.9
py4j==0.10.9.7
pycparser==2.22
PySocks==1.7.1
pyspark==3.5.1
requests==2.31.0
requests-file==2.0.0
tldextract==5.1.2
urllib3==2.2.1
Werkzeug==3.0.2
zope.event==5.0
zope.interface==6.3
